Rename the time plugin to avoid confusion with python's time module. 	- It avoids getting errors such as 	'builtin_function_or_method' object has no attribute 'sleep'
